// src/frontend/src/types/backend.ts
export type Time = bigint;

export interface Article {
  id: bigint;
  title: string;
  content: string;
  date: Time;
  imageUrl: string;
  excerpt: string;
  category: string;
}

export interface TeamMember {
  id: bigint;
  bio: string;
  name: string;
  role: string;
  imageUrl: string;
}